backend/requirements.txt
# Backend Dependencies - Job Board Solo Project

# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
greenlet==3.1.1

# Pydantic (updated for Python 3.13 support)
pydantic==2.10.3
pydantic-settings==2.7.0

# Environment
python-dotenv==1.0.1

# CORS
python-multipart==0.0.20

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# OAuth 2.0
authlib==1.3.0
httpx==0.27.0

# Rate Limiting
slowapi==0.1.9
limits==3.8.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.3
